requirements.txt
llama-cpp-python==0.3.0
huggingface-hub==0.25.2
fastapi==0.115.0
uvicorn==0.30.6